refactor: rename MinIOStorage test helpers to S3Storage
- Rename tests/remote/MinIOStorage_setup.py → S3Storage_setup.py
- Rename tests/remote/test_MinIOStorage.py → test_S3Storage.py
- Rename TmpMinIOStorage → TmpS3Storage; TestMinIOStorage → TestS3CompatibleStorage
- Update all imports in tests/fixtures.py, tests/cli/fixtures.py, tests/e2e/conftest.py
