+++ tests/cli/fixtures.py
@@ -2,7 +2,7 @@ import sys
 
 import pytest
 
-from tests.remote.MinIOStorage_setup import MinIOSetup, TmpMinIOStorage
+from tests.remote.S3Storage_setup import MinIOSetup, TmpS3Storage as TmpMinIOStorage
 
 from .path import data
