+++ tests/remote/S3Storage_setup.py
@@ -9,7 +9,7 @@ from testcontainers.minio import MinioContainer
 import yaml
 
 from omnibenchmark.model import Benchmark
-from omnibenchmark.remote.S3Storage import S3CompatibleStorage as MinIOStorage
+from omnibenchmark.remote.S3Storage import S3CompatibleStorage
 from omnibenchmark.remote.RemoteStorage import StorageOptions
 
 MINIO_IMAGE = "minio/minio:RELEASE.2024-06-13T22-53-53Z"
@@ -38,7 +38,7 @@ class MinIOSetup:
         self.minio.stop()
 
 
-class TmpMinIOStorage:
+class TmpS3Storage:
     """Manages a temporary bucket for a single test."""
 
     def __init__(self, testcontainer: MinIOSetup) -> None:
@@ -94,7 +94,7 @@ class TmpMinIOStorage:
             shutil.copyfile(env_path, env_path_after)
 
     def get_storage_client(self):
-        return MinIOStorage(
+        return S3CompatibleStorage(
             auth_options=self.auth_options,
             storage_options=self.storage_options,
             benchmark=self.bucket_name,
